Tested CachedYamlLine
